Added small subsets of msmarco-passage dev/eval per #24
Since these subsets appear to only be available from the larger
collectionandqueries file, I moved the collection download to also use
this file.

Because of how the tar file was created, there's a bit of a lag when
loading dev/small and eval/small for the first time it's loaded; it
needs to run through the (much larger) document collection first. I
could use a gzip checkpointing approach (as is done for clueweb source
file lookups), but this is overkill because the files are cached after
the first access.

(Also, a minor encoding fix in the documentation for trec-mandarin)