fixed subtle bug with mr-tydi when requesting docs_count before docs_iter
Due to a conflicting file path; when docs_count is requested, it triggers
the creation of the docstore, and the docstore ensures that its parent
directory exists. The problem is that then is the same directory that
is used for the extraction of the source files, so when it checks to see
if it needs to download, it thinks the source has already been extracted.

There were several options to fix, but I decided to separate out the source
files from the docstore. Included migration.

Thanks for spotting @eugene-yang!