added --head_precheck option to test.download
Since the head http verb is inexpensive, this allows for more frequent
probing of whether files are potentially still available. If it fails,
a proper get request is issued to check the contents.

Note that a successful head request does not necessarily mean that the
downloadable content is still unchanged; a periodic full download
is still needed. However, based on the failures seen up to now, most
cases are due to moved files or inaccessible servers, both of which
should be identified with the head request. Therefore, this can be
done more frequently, so as to identify possible problems sooner.